Added config.json.dist file
The username and password of the account used should be stored in
config.json file. config.json.dist file is added to help users in
creating this file. They can rename the file to config.json

config.json file differs from user to user and should not be
version tracked.